fix (multi-platform): Multiple problems in linkEnvironmentVariable
I'm in the process of setting up jest test suites for the repo and
realized I'm in one hell of a pickle when it comes to LocalStorage
emulation in NodeJS. As my API currently stands, it's not
capable of hosting package dependent local storage objects without
merging the local storage objects from both. Which isn't at all
ideal. I'm thinking about using a custom cache extension API on
the node.js end and boostraping that. But I don't want to make
any API features platform dependant if I can help it.

Fixes:
 * Conditional type tests for linkEnviromentVariable now work properly
 * Fixed improper environment binding assignment.