checkpoint: Backing up accidental NodeJS feature.
I'm in the middle of re-writing the API to use the
standard ES5/6 (too lazy to google) `Proxy` object as a wrapper and
I think I may have accidentally implemented a temp-dir
backup save feature that has no purpose for our cache system
layout. The cache file instanciated by `PersistentIrony`
stores an object that's always loaded in memory anyway.
Due to this fact, creating a temp file for offloading
the object is unnecessary except in the irregular circumstance
that a user temporarily looses access to the disk or storage location
which their cache resides on, then regains access to it
after the application using our API shuts down; but not before
the Operating System shuts down...

I think I just remembered what that feature is for...
Catastrophic application failure recovery lol.
*Damn that is niche.*